update: moving unit tests back in da_job file
